LPD-98243 Move the unit tests into the test folder as spec files
