+++ workspaces/liferay-aihub-workspace/client-extensions/liferay-aihub-custom-element/src/test/api.spec.ts
@@ -5,7 +5,7 @@
 
 import {afterEach, beforeEach, describe, expect, it, vi} from 'vitest';
 
-import {getChatbotConfiguration} from './api';
+import {getChatbotConfiguration} from '../api';
 
 const configuration = {title: 'My Chatbot'};
